+++ lib/users/profile.dart
@@ -1,20 +1,107 @@
 import 'package:flutter/material.dart';
+import 'package:warungkuy_mobile/main.dart';
+import 'package:warungkuy_mobile/home.dart';
+import 'package:warungkuy_mobile/signup.dart';
 
-class Profile extends StatefulWidget {
-  @override
-  State<Profile> createState() => _ProfileState();
-}
-
-class _ProfileState extends State<Profile> {
+class profile extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('Profile'),
-      ),
-      body: Center(
-        child: const Text(
-          'Launcher Page',
+    return MaterialApp(
+      home: Scaffold(
+        backgroundColor: Colors.blueAccent,
+        body: SafeArea(
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: <Widget>[
+              CircleAvatar(
+                radius: 50.0,
+                backgroundImage: AssetImage('assets/img/polije.png'),
+              ),
+              Text(
+                'Putri Chandrawati Sambo',
+                style: TextStyle(
+                  fontSize: 30.0,
+                  fontFamily: 'Pacifico',
+                  fontWeight: FontWeight.bold,
+                  color: Colors.white,
+                ),
+              ),
+              Text(
+                'WarungKuy'.toString(),
+                style: TextStyle(
+                  fontSize: 20.0,
+                  fontFamily: 'SourceSansPro',
+                  color: Colors.teal.shade100,
+                  fontWeight: FontWeight.bold,
+                  letterSpacing: 2.5,
+                ),
+              ),
+              SizedBox(
+                height: 20.0,
+                width: 150,
+                child: Divider(
+                  color: Colors.teal.shade100,
+                ),
+              ),
+              InkWell(
+                child: Card(
+                  margin:
+                      EdgeInsets.symmetric(vertical: 10.0, horizontal: 25.0),
+                  child: ListTile(
+                    leading: Icon(
+                      Icons.phone,
+                      color: Colors.teal,
+                    ),
+                    title: Text(
+                      '+229 96119149',
+                      style: TextStyle(
+                          fontFamily: 'SourceSansPro',
+                          fontSize: 20,
+                          color: Colors.teal.shade900),
+                    ),
+                  ),
+                ),
+              ),
+              InkWell(
+                child: Card(
+                  margin:
+                      EdgeInsets.symmetric(vertical: 10.0, horizontal: 25.0),
+                  child: ListTile(
+                    leading: Icon(
+                      Icons.email,
+                      color: Colors.teal,
+                    ),
+                    title: Text(
+                      'fadcrepin@gmail.com',
+                      style: TextStyle(
+                          fontFamily: 'SourceSansPro',
+                          fontSize: 20,
+                          color: Colors.teal.shade900),
+                    ),
+                  ),
+                ),
+              ),
+              InkWell(
+                child: Card(
+                  margin:
+                      EdgeInsets.symmetric(vertical: 10.0, horizontal: 25.0),
+                  child: ListTile(
+                    leading: Icon(
+                      Icons.home,
+                      color: Colors.teal,
+                    ),
+                    title: Text(
+                      'Jalan R.A Kartini No. 17, Kepatihan, Jember, Kabupaten Jember, Jawa Timur',
+                      style: TextStyle(
+                          fontFamily: 'SourceSansPro',
+                          fontSize: 15,
+                          color: Colors.teal.shade900),
+                    ),
+                  ),
+                ),
+              ),
+            ],
+          ),
         ),
       ),
     );
