--- main.dart
@@ -1,84 +1,0 @@
-import ''
-import 'package:project/ui_view/login.dart';
-import 'package:project/ui_view/signup.dart';
-
-void main() => runApp(MyApp());
-
-class MyApp extends StatelessWidget {
-  // This widget is the root of your application.
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Login Regiter',
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-      ),
-      home: Login(),
-    );
-  }
-}
-
-class HomePage extends StatelessWidget {
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Colors.lightBlue,
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            Icon(
-              Icons.android,
-              color: Colors.white,
-              size: 45,
-            ),
-            SizedBox(
-              height: 200,
-            ),
-            Text("Welcome to Flutter",
-                style: TextStyle(color: Colors.white, fontSize: 22)),
-            SizedBox(
-              height: 10,
-            ),
-            Text(
-              "Get real-time updates about what",
-              style: TextStyle(color: Colors.white, fontSize: 18),
-            ),
-            Text(
-              "maters to you",
-              style: TextStyle(color: Colors.white, fontSize: 18),
-            ),
-            SizedBox(
-              height: 18,
-            ),
-            MaterialButton(
-              minWidth: 210,
-              color: Colors.white,
-              textColor: Colors.lightBlue,
-              child: Text(
-                "Sign Up",
-                style: TextStyle(fontWeight: FontWeight.bold, fontSize: 18),
-              ),
-              onPressed: () {
-                Navigator.push(
-                    context, MaterialPageRoute(builder: (context) => SignUp()));
-              },
-            ),
-            TextButton(
-              child: Text(
-                "Log In",
-                style: TextStyle(
-                    color: Colors.white,
-                    fontWeight: FontWeight.bold,
-                    fontSize: 18),
-              ),
-              onPressed: () {
-                Navigator.push(
-                    context, MaterialPageRoute(builder: (context) => Login()));
-              },
-            )
-          ],
-        ),
-      ),
-    );
-  }
-}
